requirements.txt
aistudio-sdk==0.3.6
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bce-python-sdk==0.9.45
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
easyocr==1.7.2
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
modelscope==1.29.2
mpmath==1.3.0
narwhals==2.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
openai==1.104.2
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.1
pandas==2.3.2
pillow==11.3.0
prettytable==3.16.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.8.28
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.13
